# Reimplemented `Channel` in terms of `Queue`