--- a/build.sbt
+++ b/build.sbt
@@ -188,9 +188,9 @@ lazy val core = crossProject(JVMPlatform, JSPlatform)
     libraryDependencies ++= Seq(
       "org.typelevel" %%% "cats-core" % "2.7.0",
       "org.typelevel" %%% "cats-laws" % "2.7.0" % Test,
-      "org.typelevel" %%% "cats-effect" % "3.4-387-e167c59",
-      "org.typelevel" %%% "cats-effect-laws" % "3.4-387-e167c59" % Test,
-      "org.typelevel" %%% "cats-effect-testkit" % "3.4-387-e167c59" % Test,
+      "org.typelevel" %%% "cats-effect" % "3.4-389-3862cf0",
+      "org.typelevel" %%% "cats-effect-laws" % "3.4-389-3862cf0" % Test,
+      "org.typelevel" %%% "cats-effect-testkit" % "3.4-389-3862cf0" % Test,
       "org.scodec" %%% "scodec-bits" % "1.1.30",
       "org.typelevel" %%% "scalacheck-effect-munit" % "1.0.3" % Test,
       "org.typelevel" %%% "munit-cats-effect-3" % "1.0.7" % Test,