--- a/build.sbt
+++ b/build.sbt
@@ -10,7 +10,7 @@ ThisBuild / startYear := Some(2013)
 
 val NewScala = "2.13.8"
 
-ThisBuild / crossScalaVersions := Seq("3.1.3", "2.12.17", NewScala)
+ThisBuild / crossScalaVersions := Seq("3.2.0", "2.12.17", NewScala)
 ThisBuild / tlVersionIntroduced := Map("3" -> "3.0.3")
 
 ThisBuild / githubWorkflowOSes := Seq("ubuntu-22.04")