--- a/build.sbt
+++ b/build.sbt
@@ -205,9 +205,9 @@ lazy val core = crossProject(JVMPlatform, JSPlatform, NativePlatform)
     libraryDependencies ++= Seq(
       "org.typelevel" %%% "cats-core" % "2.8.0",
       "org.typelevel" %%% "cats-laws" % "2.8.0" % Test,
-      "org.typelevel" %%% "cats-effect" % "3.4.0-RC2",
-      "org.typelevel" %%% "cats-effect-laws" % "3.4.0-RC2" % Test,
-      "org.typelevel" %%% "cats-effect-testkit" % "3.4.0-RC2" % Test,
+      "org.typelevel" %%% "cats-effect" % "3.4-7154d08",
+      "org.typelevel" %%% "cats-effect-laws" % "3.4-7154d08" % Test,
+      "org.typelevel" %%% "cats-effect-testkit" % "3.4-7154d08" % Test,
       "org.scodec" %%% "scodec-bits" % "1.1.34",
       "org.typelevel" %%% "scalacheck-effect-munit" % "2.0.0-M2" % Test,
       "org.typelevel" %%% "munit-cats-effect" % "2.0.0-M3" % Test,