~~Do not merge. Depends on a snapshot of Cats Effect 3.4 (well, technically could be implemented without the snapshot, but it's better to wait for `tryTakeN`)~~. The snapshot in question includes the (still unmerged) improvements to `Queue` on any `Async[F]` (typelevel/cats-effect#2885 and typelevel/cats-effect#2914), but critically does not include an implementation of typelevel/cats-effect#2890, which is unquestionably the main bottleneck here:

Before:

```
[info] Benchmark                              (size)   Mode  Cnt     Score     Error  Units
[info] ChannelBenchmark.sendPull                  64  thrpt   10  7348.611 ± 207.064  ops/s
[info] ChannelBenchmark.sendPull                1024  thrpt   10  3850.531 ±  10.141  ops/s
[info] ChannelBenchmark.sendPull               16384  thrpt   10   372.989 ±   1.049  ops/s
[info] ChannelBenchmark.sendPullPar8              64  thrpt   10  9122.642 ±  49.265  ops/s
[info] ChannelBenchmark.sendPullPar8            1024  thrpt   10  3558.006 ±   5.951  ops/s
[info] ChannelBenchmark.sendPullPar8           16384  thrpt   10   254.749 ±   0.650  ops/s
[info] ChannelBenchmark.sendPullParUnlimited      64  thrpt   10  9367.322 ±  25.509  ops/s
[info] ChannelBenchmark.sendPullParUnlimited    1024  thrpt   10   897.318 ±  17.716  ops/s
[info] ChannelBenchmark.sendPullParUnlimited   16384  thrpt   10    39.973 ±   0.367  ops/s
```

After:

```
[info] Benchmark                              (size)   Mode  Cnt     Score    Error  Units
[info] ChannelBenchmark.sendPull                  64  thrpt   10  5383.059 ± 42.282  ops/s
[info] ChannelBenchmark.sendPull                1024  thrpt   10  2564.141 ±  5.962  ops/s
[info] ChannelBenchmark.sendPull               16384  thrpt   10   254.210 ±  0.407  ops/s
[info] ChannelBenchmark.sendPullPar8              64  thrpt   10  4840.209 ± 21.118  ops/s
[info] ChannelBenchmark.sendPullPar8            1024  thrpt   10  2427.766 ±  2.502  ops/s
[info] ChannelBenchmark.sendPullPar8           16384  thrpt   10   179.873 ±  0.524  ops/s
[info] ChannelBenchmark.sendPullParUnlimited      64  thrpt   10  7256.407 ± 29.061  ops/s
[info] ChannelBenchmark.sendPullParUnlimited    1024  thrpt   10  1334.681 ±  6.618  ops/s
[info] ChannelBenchmark.sendPullParUnlimited   16384  thrpt   10    77.232 ±  1.446  ops/s
```

To me, the fact that this is even *close* to the hand-rolled version is pretty cool, and a slightly less naive `stream` would probably put it almost on-par even without the optimized `tryTakeN`. Once we get the latter though this should be head and shoulders above the prior implementation.

Closes #2852