@armanbilge 2022-09-30
```
Error: Exception in thread "io-compute-4" java.lang.OutOfMemoryError: Requested array size exceeds VM limit
	at cats.effect.std.Queue$UnsafeBounded.<init>(Queue.scala:770)
	at cats.effect.std.Queue$BoundedAsyncQueue.<init>(Queue.scala:412)
	at cats.effect.std.Queue$.boundedForAsync$$anonfun$1(Queue.scala:96)
	at cats.effect.std.Queue$$$Lambda$1222/0x0000000800f5d398.apply(Unknown Source)
	at cats.effect.IOFiber.runLoop(IOFiber.scala:401)
	at cats.effect.IOFiber.execR(IOFiber.scala:1309)
	at cats.effect.IOFiber.run(IOFiber.scala:118)
	at cats.effect.unsafe.WorkerThread.run(WorkerThread.scala:580)
```