@armanbilge 2022-09-30
So it seems like anyone who was using `bounded` with a large number and expecting it to be allocated lazily is in for a bad time 🙃