@djspiewak 2022-09-30
> So it seems like anyone who was using bounded with a large number and expecting it to be allocated lazily is in for a bad time 🙃

Yes. They shouldn't do that. :-P I mean, we can special case large bounds, but genuinely this just isn't a case which should be supported. It only worked before by coincidence.