@djspiewak 2022-10-09
Getting better!

Not quite better across the board yet, but doing really well. Because of how things are optimized, I'm pretty sure that the new implementation gets better the faster the producers run relative to the consumer, though the numbers don't seem to intuitively bear that out. Still fiddling with things, and the upstream `Queue` has plenty of room for improvement (especially `unbounded`).

## Before

```
[info] Benchmark                              (size)   Mode  Cnt     Score     Error  Units
[info] ChannelBenchmark.sendPull                  64  thrpt   10  7735.063 ±  33.841  ops/s
[info] ChannelBenchmark.sendPull                1024  thrpt   10  3760.871 ±   9.716  ops/s
[info] ChannelBenchmark.sendPull               16384  thrpt   10   359.674 ±   1.309  ops/s
[info] ChannelBenchmark.sendPullPar8              64  thrpt   10  8533.059 ± 182.194  ops/s
[info] ChannelBenchmark.sendPullPar8            1024  thrpt   10  3464.678 ±   8.370  ops/s
[info] ChannelBenchmark.sendPullPar8           16384  thrpt   10   247.827 ±   2.399  ops/s
[info] ChannelBenchmark.sendPullParUnlimited      64  thrpt   10  8921.664 ±  27.248  ops/s
[info] ChannelBenchmark.sendPullParUnlimited    1024  thrpt   10  1002.150 ±  26.415  ops/s
[info] ChannelBenchmark.sendPullParUnlimited   16384  thrpt   10    39.949 ±   0.363  ops/s
```

## After

```
[info] Benchmark                              (size)   Mode  Cnt     Score     Error  Units
[info] ChannelBenchmark.sendPull                  64  thrpt   10  8092.902 ± 112.613  ops/s
[info] ChannelBenchmark.sendPull                1024  thrpt   10  1934.434 ±  13.293  ops/s
[info] ChannelBenchmark.sendPull               16384  thrpt   10   144.031 ±   2.505  ops/s
[info] ChannelBenchmark.sendPullPar8              64  thrpt   10  7045.267 ±  37.947  ops/s
[info] ChannelBenchmark.sendPullPar8            1024  thrpt   10  2515.080 ±  13.203  ops/s
[info] ChannelBenchmark.sendPullPar8           16384  thrpt   10   225.750 ±   1.203  ops/s
[info] ChannelBenchmark.sendPullParUnlimited      64  thrpt   10  7241.213 ± 106.143  ops/s
[info] ChannelBenchmark.sendPullParUnlimited    1024  thrpt   10  1468.275 ±  20.139  ops/s
[info] ChannelBenchmark.sendPullParUnlimited   16384  thrpt   10    76.752 ±  10.437  ops/s
```