@djspiewak 2022-10-09
…and a little better:

```
[info] Benchmark                              (size)   Mode  Cnt     Score     Error  Units
[info] ChannelBenchmark.sendPull                  64  thrpt   10  8847.457 ±  77.061  ops/s
[info] ChannelBenchmark.sendPull                1024  thrpt   10  1870.886 ±  16.247  ops/s
[info] ChannelBenchmark.sendPull               16384  thrpt   10   132.900 ±   0.244  ops/s
[info] ChannelBenchmark.sendPullPar8              64  thrpt   10  8171.220 ± 166.752  ops/s
[info] ChannelBenchmark.sendPullPar8            1024  thrpt   10  2673.881 ±  26.653  ops/s
[info] ChannelBenchmark.sendPullPar8           16384  thrpt   10   218.006 ±   1.377  ops/s
[info] ChannelBenchmark.sendPullParUnlimited      64  thrpt   10  8604.675 ± 247.214  ops/s
[info] ChannelBenchmark.sendPullParUnlimited    1024  thrpt   10  1415.921 ±  18.295  ops/s
[info] ChannelBenchmark.sendPullParUnlimited   16384  thrpt   10    79.396 ±   2.864  ops/s
```