@djspiewak 2022-10-26
@SystemFw I think this is ready for another look. Your `Ref` consolidation advice is very good, and probably actually improved the performance of `send` (I'm checking now). I also corrected the issue with `closed` being fulfilled after quiescence rather than upon closure. Additionally, I fixed a bug in CE's `Queue.synchronous` which was the origin of the broadcast hang.

Depends on a hash snapshot for the time being, hence why this is marked as draft.