@djspiewak 2022-10-26
Updated with benchmark results. The simple cases got faster with the consolidated `Ref`, but the more complex cases got *much* slower, and any case involving contention also got noticeably slower:

## `main`

```
[info] Benchmark                              (size)   Mode  Cnt     Score     Error  Units
[info] ChannelBenchmark.sendPull                  64  thrpt   10  6893.574 ± 125.608  ops/s
[info] ChannelBenchmark.sendPull                1024  thrpt   10  3535.811 ±  39.588  ops/s
[info] ChannelBenchmark.sendPull               16384  thrpt   10   361.574 ±   1.307  ops/s
[info] ChannelBenchmark.sendPullPar8              64  thrpt   10  8176.797 ± 118.350  ops/s
[info] ChannelBenchmark.sendPullPar8            1024  thrpt   10  3407.982 ±  22.908  ops/s
[info] ChannelBenchmark.sendPullPar8           16384  thrpt   10   250.676 ±   1.205  ops/s
[info] ChannelBenchmark.sendPullParUnlimited      64  thrpt   10  8622.935 ±  48.615  ops/s
[info] ChannelBenchmark.sendPullParUnlimited    1024  thrpt   10   925.061 ±  27.972  ops/s
[info] ChannelBenchmark.sendPullParUnlimited   16384  thrpt   10    39.134 ±   0.659  ops/s
```

## This PR

```
[info] Benchmark                              (size)   Mode  Cnt     Score     Error  Units
[info] ChannelBenchmark.sendPull                  64  thrpt   10  9343.856 ±  42.086  ops/s
[info] ChannelBenchmark.sendPull                1024  thrpt   10  2161.634 ±  12.066  ops/s
[info] ChannelBenchmark.sendPull               16384  thrpt   10    22.743 ±   0.856  ops/s
[info] ChannelBenchmark.sendPullPar8              64  thrpt   10  7830.841 ±  41.508  ops/s
[info] ChannelBenchmark.sendPullPar8            1024  thrpt   10  2175.548 ±  56.683  ops/s
[info] ChannelBenchmark.sendPullPar8           16384  thrpt   10   116.707 ±   2.704  ops/s
[info] ChannelBenchmark.sendPullParUnlimited      64  thrpt   10  8125.314 ± 164.499  ops/s
[info] ChannelBenchmark.sendPullParUnlimited    1024  thrpt   10  1375.947 ±  33.510  ops/s
[info] ChannelBenchmark.sendPullParUnlimited   16384  thrpt   10    54.626 ±   1.014  ops/s
```